mini_projeto_api_autenticacao/pyproject.toml
[tool.poetry]
name = "mini-projeto-api-autenticacao"
version = "0.1.0"
description = ""
authors = ["dolivmorais <dolivmorais@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
streamlit = "^1.37.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
